# Endpoint for listing allowed actions on Multiple Resources