--- a/lib/core/config/config.auth.js
+++ b/lib/core/config/config.auth.js
@@ -63,7 +63,8 @@ const Actions = {
 
   // authorization
   CheckAccess: 'authorization:authn:access',
-  ListActions: 'authorization:authn:actions'
+  ListActions: 'authorization:authn:actions',
+  ListActionsOnResources: 'authorization:authn:resources:actions'
 }
 
 module.exports = {