--- a/lib/core/index.js
+++ b/lib/core/index.js
@@ -13,7 +13,8 @@ module.exports = {
 
   authorize: {
     isUserAuthorized: authorizeOps.isUserAuthorized,
-    listActions: authorizeOps.listAuthorizations
+    listActions: authorizeOps.listAuthorizations,
+    listAuthorizationsOnResources: authorizeOps.listAuthorizationsOnResources
   },
 
   organizations: {