--- a/lib/core/lib/ops/authorizeOps.js
+++ b/lib/core/lib/ops/authorizeOps.js
@@ -62,10 +62,31 @@ const authorize = {
     ], function (err, actions) {
       cb(err, { actions })
     })
+  },
+
+  /**
+   * List all user's actions on a given list of resources
+   *
+   * @param  {Object}   options { userId, resources }
+   * @param  {Function} cb
+   */
+  listAuthorizationsOnResources: function listAuthorizationsOnResources ({ userId, resources, organizationId }, cb) {
+    async.waterfall([
+      function validate (next) {
+        Joi.validate({ userId, resources, organizationId }, validationRules.listAuthorizationsOnResources, badRequestWrap(next))
+      },
+      function listPolicies (result, next) {
+        policyOps.listAllUserPolicies({ userId, organizationId }, badImplementationWrap(next))
+      },
+      function listAuthorizationsOnResources (policies, next) {
+        iam(policies).actionsOnResources({ resources }, badImplementationWrap(next))
+      }
+    ], cb)
   }
 }
 
 authorize.isUserAuthorized.validate = validationRules.isUserAuthorized
 authorize.listAuthorizations.validate = validationRules.listAuthorizations
+authorize.listAuthorizationsOnResources.validate = validationRules.listAuthorizationsOnResources
 
 module.exports = authorize