--- a/lib/core/lib/ops/authorizeOps.js
+++ b/lib/core/lib/ops/authorizeOps.js
@@ -35,7 +35,7 @@ const authorize = {
         policyOps.listAllUserPolicies({ userId, organizationId }, badImplementationWrap(next))
       },
       function check (policies, next) {
-        iam(policies).isAuthorized({ resource, action }, badImplementationWrap(next))
+        next(null, iam(policies).isAuthorized({ resource, action }))
       }
     ], function (err, access) {
       cb(err, { access })