--- a/lib/core/lib/ops/iam.js
+++ b/lib/core/lib/ops/iam.js
@@ -42,8 +42,48 @@ module.exports = function (policies) {
     }
   }
 
+  function actionsOnResources ({ resources }, done) {
+    try {
+      const resultMap = {}
+
+      const statements = _(policies).reduce((statements, policy) => {
+        return _.concat(statements, policy.Statement)
+      }, [])
+
+      resources.forEach((resource) => {
+        resultMap[resource] = []
+        statements.forEach((statement) => {
+          if (statement.Effect === 'Allow') {
+            statement.Resource.forEach((r) => {
+              if (pbac.conditions.StringLike(r, resource)) {
+                statement.Action.forEach((action) => {
+                  isAuthorized({resource, action}, (err, access) => {
+                    if (err) return
+                    if (access) resultMap[resource].push(action)
+                  })
+                })
+              }
+            })
+          }
+        })
+        resultMap[resource] = _.uniq(resultMap[resource])
+      })
+
+      let result = _(resultMap)
+        .reduce((result, actions, resource) => {
+          result.push({resource, actions})
+          return result
+        }, [])
+
+      return done(null, result)
+    } catch (e) {
+      return done(e)
+    }
+  }
+
   return {
     isAuthorized: isAuthorized,
-    actions: actions
+    actions: actions,
+    actionsOnResources: actionsOnResources
   }
 }