--- a/lib/core/lib/ops/iam.js
+++ b/lib/core/lib/ops/iam.js
@@ -7,7 +7,7 @@ module.exports = function (policies) {
   const pbac = new PBAC(policies)
 
   function isAuthorized (params, done) {
-    return done(null, pbac.evaluate(params))
+    return pbac.evaluate(params)
   }
 
   function actions ({ resource }, done) {
@@ -21,10 +21,9 @@ module.exports = function (policies) {
           statement.Resource.forEach((r) => {
             if (pbac.conditions.StringLike(r, resource)) {
               statement.Action.forEach((action) => {
-                isAuthorized({ resource, action }, (err, access) => {
-                  if (err) return
-                  if (access) actions.push(action)
-                })
+                if (isAuthorized({resource, action})) {
+                  actions.push(action)
+                }
               })
             }
           })
@@ -57,10 +56,9 @@ module.exports = function (policies) {
             statement.Resource.forEach((r) => {
               if (pbac.conditions.StringLike(r, resource)) {
                 statement.Action.forEach((action) => {
-                  isAuthorized({resource, action}, (err, access) => {
-                    if (err) return
-                    if (access) resultMap[resource].push(action)
-                  })
+                  if (isAuthorized({resource, action})) {
+                    resultMap[resource].push(action)
+                  }
                 })
               }
             })