--- a/lib/core/lib/ops/validation.js
+++ b/lib/core/lib/ops/validation.js
@@ -30,6 +30,7 @@ const validationRules = {
   users: Joi.array().required().items(requiredString).description('User IDs'),
   policies: Joi.array().required().items(requiredString).description('Policies IDs'),
   teams: Joi.array().required().items(requiredString).description('Teams IDs'),
+  resources: Joi.array().max(10).items(requiredString.description('A single resource')).single().required().description('A list of Resources'),
 
   statements: Joi.object({
     Statement: Joi.array().items(Joi.object({
@@ -251,6 +252,11 @@ const authorize = {
   listAuthorizations: {
     userId: validationRules.userId,
     resource: requiredString.description('The resource that the user wants to perform the action on'),
+    organizationId: validationRules.organizationId
+  },
+  listAuthorizationsOnResources: {
+    userId: validationRules.userId,
+    resources: validationRules.resources,
     organizationId: validationRules.organizationId
   }
 }