--- a/lib/plugin/routes/public/authorization.js
+++ b/lib/plugin/routes/public/authorization.js
@@ -75,6 +75,40 @@ exports.register = function (server, options, next) {
     }
   })
 
+  server.route({
+    method: 'GET',
+    path: '/authorization/list/{userId}',
+    handler: function (request, reply) {
+      const { organizationId } = request.udaru
+      const { userId } = request.params
+      const { resources } = request.query
+      const params = {
+        userId,
+        resources,
+        organizationId
+      }
+
+      udaru.authorize.listAuthorizationsOnResources(params, reply)
+    },
+    config: {
+      plugins: {
+        auth: {
+          action: Action.ListActionsOnResources,
+          resource: 'authorization/actions/resources'
+        }
+      },
+      validate: {
+        params: _.pick(udaru.authorize.listAuthorizationsOnResources.validate, ['userId']),
+        query: _.pick(udaru.authorize.listAuthorizationsOnResources.validate, ['resources']),
+        headers
+      },
+      description: 'List all the actions a user can perform on a list of resources',
+      notes: 'The GET /authorization/list/{userId} endpoint returns a list of all the actions a user\ncan perform on a given list of resources.\n',
+      tags: ['api', 'authorization'],
+      response: {schema: swagger.UserActionsOnResources}
+    }
+  })
+
   next()
 }