--- a/lib/plugin/swagger.js
+++ b/lib/plugin/swagger.js
@@ -83,10 +83,16 @@ const UserActions = Joi.object({
   actions: Joi.array().items(Joi.string())
 }).label('UserActions')
 
+const UserActionsOnResources = Joi.array().items(Joi.object({
+  resource: Joi.string(),
+  actions: Joi.array().items(Joi.string())
+})).label('UserActionsOnResources')
+
 module.exports = {
   List,
   User,
   UserActions,
+  UserActionsOnResources,
   Team,
   Policy,
   PolicyRef,