This new endpoint allows us to list the actions a user can perform on a list of resources

### Summary
When calling, the endpoint looks like this:

```
GET /authorization/list/some-user-id?resources=foo&resources=bar&resources=baz
```

which will return a list like this:

```
 [
  {
    resource: 'foo',
    actions: ['Read']
  },
  {
    resource: 'bar',
    actions: ['*']
  },
  {
    resource: 'baz',
    actions: []
  }
]
```

__Small Note on Validation__
The endpoint will still work if only one resource is passed in the querystring.

The Joi schema for resources uses ['Array.single()`](https://github.com/hapijs/joi/blob/v10.6.0/API.md#arraysingleenabled) which will optionally accept a single value and put it in an array. e.g.

```
GET /authorization/list/some-user-id?resources=foo
```

will result in `request.query.resources === ['foo']`

### How to Test

Assuming a fresh install, load in the test data with

```
npm run pg:init-test-db
```

and then we can test the endpoint using some of the data inserted by the `fixtures.sql` script with the following:

```
curl -X GET --header 'Accept: application/json' --header 'authorization: ROOTid' --header 'org: WONKA' 'http://localhost:8080/authorization/list/ManyPoliciesId?resources=/myapp/users/dara&resources=/myapp/documents/no_access&resources=/myapp/teams/*'
```

which will give back:

```
[
  {
    "resource": "/myapp/users/dara",
    "actions": [
      "Read"
    ]
  },
  {
    "resource": "/myapp/documents/no_access",
    "actions": []
  },
  {
    "resource": "/myapp/teams/*",
    "actions": [
      "Read",
      "Delete",
      "Edit"
    ]
  }
]
```