@darahayes 2017-08-09
@p16 yeah come to think of it, the core method is a bit lazy. It's going to list all policies for every resource. Bad...

I'll should be able to reimplement the core method fairly quickly tomorrow morning