@paolochiodi 2017-08-10
I'd like to report here, for the records, the concern that I had about the list endpoint since we first started discussing it at the beginning of the project.

I find the endpoint as currently implemented (either before or after the PR) a bit unuseful and misleading, because it's not actually returning the list of actions that can be performed, but rather the list of action "rules" that have been inserted for the user.

In the example reported in this PR:
```js
 [
  {
    resource: 'foo',
    actions: ['Read']
  },
  {
    resource: 'bar',
    actions: ['*']
  },
  {
    resource: 'baz',
    actions: []
  }
]
```

we are not telling which actions the user can perform on resource bar, we are just telling that it can do everything. The problem is more visible if we return something like 

```
actions: [
 "authorization:teams:*"
]
```

In these cases the clients needs to re-implment a part of udaru (the pattern matching) in order to get the answer it wants.

I'd suggest to also require a list of actions as input, we could then return a matrix with allow/deny for each resource/action combination. 
Again, following the previous example:

```
GET /authorization/list/some-user-id?resources=foo&resources=bar&resources=ba&actions=Read&actions=Write
```

would result in

```json
{
  "foo": {
    "Read" : "allow",
    "Write": "deny"
  },
  "bar": {
    "Read" : "allow",
    "Write": "allow"
  },
  "baz": {
    "Read" : "deny",
    "Write": "deny"
  }
}
```

I find this solution easier to comprehend and to communicate to the client while still covering 99% of the use cases of the list endpoint