@feugy 2017-08-10
@paolochiodi, considering how we used the two main endpoints (access and list) in our project, your suggestion would be more a refinement of access than list.

Our UI is using list endpoint to list all possible action on a given resource, without prior knowledge of the action list.
Our backend is using access endpoint to check whether a given action is allowed on given resource.

What you're suggesting is really promising for us, because we often call access 3/4 times with different actions and resources during the same operation. This would be a significant improvement.

However I still think there's a room for an endpoint listing allowed actions without knowing them in advance.