@darahayes 2017-08-10
Very interesting and valid points guys. Based on what I'm reading here, I believe the functionality @paolochiodi says he would like to see is more of an extension of the access functionality as opposed to the list functionality.

In the original issue #390 we mentioned doing two endpoints. One that provides the functionality in this PR and another almost as Paolo described, although the idea wasn't fleshed out very well. Based off both of your comments there is definitely scope for another endpoint. I can get this done in a separate PR.