@darahayes 2017-08-10
Okay @p16 I've updated moved the implementation into the `iam.js` script. This has made the implementation about 3-4 times faster than the previous one.

Two things to note:

1. The code is not as elegant as the other `actions` function. I found that chaining a bunch of `filter` and `map` functions made it pretty slow. Instead I wrote it in a more elementary style. I hope it's still readable.
2. I found that there's no need to actually check `isAuthorized({action, resource})` because you already know that the action and resource is inside a statement with `Effect: Allow` If we know `Effect === 'Allow'` then it must be authorized right? Please correct me if I'm wrong but I seem to be getting the desired results.

### Benchmarking
I also added benchmarking tests for the original list endpoint and the new one. 

The new endpoint is about 3-4 times faster than my first implementation. Also based off the tests I've done, the new multi resource endpoint is as fast (and sometimes faster) than the original single list endpoint. (10 resources vs 1 resource).

Please let me know what you think.

### Old Implementation of Multi Resource List Endpoint

```
Stat         Avg     Stdev   Max
Latency (ms) 8718.61 728.34  9975
Req/Sec      10.3    16.02   42
Bytes/Sec    9.11 kB 14.2 kB 38.9 kB

103 requests in 10s, 90.5 kB read
397 errors (394 timeouts)
Detailed Result:

 { title: 'GET authorization/list/{userId}',
  url: 'http://localhost:8080/authorization/list/ManyPoliciesId?resources=/myapp/users/filippo&resources=/myapp/documents/no_access&resources=/myapp/teams/foo&resources=/myapp/teams/foo1&resources=/myapp/teams/foo2&resources=/myapp/teams/foo3&resources=/myapp/teams/foo4&resources=/myapp/teams/foo5&resources=/myapp/teams/foo6&resources=/myapp/teams/foo7',
  requests:
   { average: 10.3,
     mean: 10.3,
     stddev: 16.02,
     min: 0,
     max: 42,
     total: 103,
     sent: 1000 },
  latency:
   { average: 8718.61,
     mean: 8718.61,
     stddev: 728.34,
     min: 7449,
     max: 9975,
     p50: 8719,
     p75: 9311,
     p90: 9731,
     p99: 9967,
     p999: 9975,
     p9999: 9975,
     p99999: 9975 },
  throughput:
   { average: 9113.6,
     mean: 9113.6,
     stddev: 14223.66,
     min: 0,
     max: 38911,
     total: 90537 },
  errors: 397,
  timeouts: 394,
  duration: 10,
  start: 2017-08-10T14:42:55.442Z,
  finish: 2017-08-10T14:43:05.763Z,
  connections: 500,
  pipelining: 1,
  non2xx: 0,
  '1xx': 0,
  '2xx': 103,
  '3xx': 0,
  '4xx': 0,
  '5xx': 0 }
```
### New Implementation of Multi Resource List Endpoint

Whereas with the newer implementation I got the following:

```
Stat         Avg     Stdev   Max
Latency (ms) 4685.39 563.64  5361
Req/Sec      92.4    102.96  276
Bytes/Sec    80.8 kB 90.2 kB 246 kB

924 requests in 10s, 812 kB read
Detailed Result:

 { title: 'GET authorization/list/{userId}',
  url: 'http://localhost:8080/authorization/list/ManyPoliciesId?resources=/myapp/users/filippo&resources=/myapp/documents/no_access&resources=/myapp/teams/foo&resources=/myapp/teams/foo1&resources=/myapp/teams/foo2&resources=/myapp/teams/foo3&resources=/myapp/teams/foo4&resources=/myapp/teams/foo5&resources=/myapp/teams/foo6&resources=/myapp/teams/foo7',
  requests:
   { average: 92.4,
     mean: 92.4,
     stddev: 102.96,
     min: 0,
     max: 276,
     total: 924,
     sent: 1424 },
  latency:
   { average: 4685.39,
     mean: 4685.39,
     stddev: 563.64,
     min: 2232,
     max: 5361,
     p50: 5012,
     p75: 5086,
     p90: 5128,
     p99: 5327,
     p999: 5358,
     p9999: 5361,
     p99999: 5361 },
  throughput:
   { average: 80844.8,
     mean: 80844.8,
     stddev: 90217.89,
     min: 0,
     max: 245759,
     total: 812196 },
  errors: 0,
  timeouts: 0,
  duration: 10,
  start: 2017-08-10T14:43:41.516Z,
  finish: 2017-08-10T14:43:51.679Z,
  connections: 500,
  pipelining: 1,
  non2xx: 0,
  '1xx': 0,
  '2xx': 924,
  '3xx': 0,
  '4xx': 0,
  '5xx': 0 }
```

### Single Resource List Endpoint

```
Running 10s test @ http://localhost:8080/authorization/list/ROOTid/resource_a
500 connections

Stat         Avg     Stdev   Max
Latency (ms) 5267.52 488.04  6196
Req/Sec      61.3    90.38   260
Bytes/Sec    14.8 kB 21.7 kB 63.5 kB

613 requests in 10s, 148 kB read
83 errors (0 timeouts)
Detailed Result:

 { title: 'GET authorization/list/{userId}/{resource}',
  url: 'http://localhost:8080/authorization/list/ROOTid/resource_a',
  requests:
   { average: 61.3,
     mean: 61.3,
     stddev: 90.38,
     min: 0,
     max: 260,
     total: 613,
     sent: 1196 },
  latency:
   { average: 5267.52,
     mean: 5267.52,
     stddev: 488.04,
     min: 4273,
     max: 6196,
     p50: 5333,
     p75: 5560,
     p90: 5923,
     p99: 6179,
     p999: 6193,
     p9999: 6196,
     p99999: 6196 },
  throughput:
   { average: 14771.2,
     mean: 14771.2,
     stddev: 21735.61,
     min: 0,
     max: 63487,
     total: 147733 },
  errors: 83,
  timeouts: 0,
  duration: 10,
  start: 2017-08-10T15:06:28.755Z,
  finish: 2017-08-10T15:06:38.931Z,
  connections: 500,
  pipelining: 1,
  non2xx: 0,
  '1xx': 0,
  '2xx': 613,
  '3xx': 0,
  '4xx': 0,
  '5xx': 0 }
```